requirements.txt
watchdog
flask
configparser
requests
python-dotenv
slack_sdk
